Fix affine16 and hoister minor graphics things

+++ emu/nds/utils/floats.go
@@ -58,7 +58,7 @@ func ConvertFromFloat4_0_12(f float32) uint16 {
 }
 
 func FloatRound(v, step float32) float32 {
-	return float32(math.Round(float64(v/step)) * float64(step))
+	return float32(math.Round(float64(v/step))) * step
 }
 
 func FloatFloor(v, step float32) float32 {
